+++ PL/Controllers/FilesController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.IO;
 using System.Linq;
 using System.Net;
@@ -41,7 +42,7 @@ namespace PL.Controllers
         }
         
         // GET api/Files
-        //[Authorize (Roles = "admin, user")]
+        [Authorize (Roles = "admin")]
         [HttpGet]
         public async Task<IActionResult> GetAllFiles()
         {
@@ -51,33 +52,58 @@ namespace PL.Controllers
         }
         
         // GET api/Files/5
-       // [Authorize (Roles = "admin, user")]
+        [Authorize (Roles = "admin, user")]
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetFileById(int id)
         {
-            var file = await _fileService.GetByIdAsync(id);
+           
+            var user = User.Identity?.Name;
+             if (user == null) throw new NullReferenceException(nameof(user));
+
+             /*if (await _userService.CheckBan(user, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }*/
+            var file = await _fileService.GetByIdAsync(id, user);
+            if (file.Id == 0)
+            {
+                return BadRequest();
+            }
             var fileModel = _mapper.Map<FileInfoModel>(file);
 
             return Ok(fileModel);
         }
         
         // GET api/Files/fileSearch/word
-       // [Authorize (Roles = "admin")]
+        [Authorize (Roles = "admin, user")]
         [HttpGet("filesSearch/{keyword}")]
-        public IActionResult GetFilesByKeyword(string keyword)
+        public async Task<IActionResult> GetFilesByKeyword(string keyword)
         {
-            var files = _fileService.GetByKeyword(keyword);
+           
+            
+            var userEmail = User.Identity?.Name;
+            
+            if (await _userService.CheckBan(userEmail, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }
+            var files = await _fileService.GetByKeyword(keyword, userEmail);
 
             return Ok(files);
         }
         
         
         // POST api/Files/upload
-       // [Authorize(Roles = "user, admin")]
+        [Authorize(Roles = "user, admin")]
         [HttpPost("upload")]
         [FileUploadOperation.FileContentType]
         public async Task<IActionResult> UploadFile([FromForm] CreateFileModel model, IFormFile uploadedFile)
         {
+            var user = User.Identity?.Name;
+            if (await _userService.CheckBan(user, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }
             
             var file = uploadedFile.OpenReadStream();
             var fileInfo = _mapper.Map<FileDto>(model);
@@ -87,7 +113,7 @@ namespace PL.Controllers
             /*var result = await Request.ReadFormAsync();
             var file = result.Files[0].OpenReadStream();*/
             var createdFile = await _fileService.AddAsync(file, fileInfo);
-           
+
             return Created(fileInfo.Title, createdFile.Url);
         }
         
@@ -96,7 +122,12 @@ namespace PL.Controllers
         public async Task<IActionResult> DownloadFile( int id)
         {
 
-            var file = await _fileService.GetByIdAsync(id);
+            var user = User.Identity?.Name;
+            if (await _userService.CheckBan(user, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }
+            var file = await _fileService.GetByIdAsync(id, user);
             if (file == null) return NotFound();
             
             var provider = new FileExtensionContentTypeProvider();
@@ -114,12 +145,32 @@ namespace PL.Controllers
         
         //!!!!!!!!!!!!!!!!!!!!!!!!
         // PUT api/Files/edit
-       // [Authorize (Roles = "admin, user")]
+        [Authorize (Roles = "admin, user")]
         [HttpPut("edit/{id:int}")]
         [FileUploadOperation.FileContentType]
         public async Task<IActionResult> EditFile( [FromForm] CreateFileModel model, IFormFile uploadedFile, int id)
         {
-            var file = uploadedFile.OpenReadStream();
+            var user = User.Identity?.Name;
+            if (await _userService.CheckBan(user, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }
+            
+                var file = uploadedFile.OpenReadStream();
+
+
+                /*var oldFile = await _fileService.GetByIdAsync(id, User.Identity.Name);
+                if (oldFile == null) return NotFound();
+                
+                var provider = new FileExtensionContentTypeProvider();
+                
+                if (!provider.TryGetContentType(file.Url, out var contentType))
+                {
+                    contentType = "application/octet-stream";
+                }
+                
+                var olddata = await _fileService.ReadFileAsync(oldFile);
+                var oldFileS = new FileStream(oldFile.Url, FileAccess.Read);*/
 
             var fileInfo = _mapper.Map<FileDto>(model);
             fileInfo.Id = id;
@@ -136,10 +187,15 @@ namespace PL.Controllers
         }
         
         // DELETE api/Files/delete/5
-       // [Authorize(Roles = "user, admin")]
+        [Authorize(Roles = "user, admin")]
         [HttpDelete("delete/{id:int}")]
         public async Task<IActionResult> DeleteFile(int id)
         {
+            var user = User.Identity?.Name;
+            if (await _userService.CheckBan(user, DateTime.Now))
+            {
+                return BadRequest("User is banned!");
+            }
             await _fileService.DeleteByIdAsync(id);
             return Ok();
         }
@@ -149,6 +205,28 @@ namespace PL.Controllers
         public async Task<IActionResult> GetFilesAccesses()
         {
             var files =_mapper.Map<IEnumerable<AccessListModel>>(await _fileService.GetFileAccesses());
+
+            return Ok(files);
+        }
+        /*
+        // GET api/Admin/users/4/files/2
+        [HttpGet("{userId:int}/files/{fileId:int}")]
+        public async Task<IActionResult> GetUserFilesById(int userId, int fileId)
+        {
+            var files = await _userService.GetAllUsersFiles(userId);
+            var file = files.FirstOrDefault(f => f.Id == fileId);
+
+            return Ok(file);
+        }
+        */
+        
+        // GET api/Files/users/4/files
+        [Authorize(Roles = "user, admin")]
+        [HttpGet("userFiles")]
+        public async Task<IActionResult> GetUserFiles()
+        {
+            var user = User.Identity.Name;
+            var files = await _fileService.GetAllUsersFiles(user);
 
             return Ok(files);
         }
